requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
ansible==9.8.0
ansible-core==2.16.3
anyio==4.11.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==25.4.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
blinker==1.4
bs4==0.0.2
certifi==2025.7.14
chardet==4.0.0
charset-normalizer==3.4.4
click==8.0.3
# cloud-init==25.1.4
colorama==0.4.4
# command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cssselect==1.3.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
dotenv==0.9.9
exceptiongroup==1.3.1
fake-useragent==2.2.0
feedparser==6.0.10
frozenlist==1.7.0
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.3
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
influxdb-client==1.49.0
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.12.0
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==23.5.0
langchain==1.1.0
langchain-core==1.1.0
langchain-mcp-adapters==0.1.14
langchain-openai==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
loguru==0.7.2
lxml==6.0.0
MarkupSafe==2.0.1
# Editable install with no version control (matain-agent==0.1.0)
-e /root/huhu/agent/matain_agent
mcp==1.22.0
more-itertools==8.10.0
multidict==6.6.3
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parse==1.20.2
pexpect==4.8.0
propcache==0.3.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==11.1.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.10.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyppeteer==2.0.0
pyquery==2.0.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-magic==0.4.24
python-multipart==0.0.20
pytz==2022.1
PyYAML==6.0.1
reactivex==4.0.4
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-html==0.10.0
requests-toolbelt==1.0.0
resolvelib==0.8.1
rpds-py==0.30.0
SecretStorage==3.3.1
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sos==4.5.6
soupsieve==2.7
sse-starlette==3.0.3
ssh-import-id==5.11
starlette==0.50.0
systemd-python==234
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
Twisted==22.1.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.38.0
w3lib==2.3.1
wadllib==1.3.6
websockets==10.4
xkit==0.0.0
xxhash==3.6.0
yarl==1.20.1
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.25.0
